refactor: split convex/characters.ts into characters / combat / items
Per docs/plans/in-progress.md, the file held three unrelated domains
(CRUD, combat/zone, items/bag) and was on track to hit 1500+ lines as
stash/vendor/skills land. Splitting now so future features have a clear
home and merge conflicts stay narrow.

New layout:

- `convex/characters.ts` — CRUD + roster only: `list`, `byId`, `create`,
  `remove`, plus private `normalize`/`isKnownClassId`/`defaultStarterWeapon`.
  Dropped from 880 to 161 lines.
- `convex/combat.ts` — `enterZone`, `enterCity`, `recordKill`, `syncHp`,
  `usePotion`, `respawnDead`. The "what's the player doing in/around a
  zone" mutations.
- `convex/items.ts` — `equipItem`, `unequipItem`, `reorderInventory`,
  `pickFromBag`, `discardFromBag`, `exitZone`, plus queries `zoneBag`,
  `inventory`, `equipped`. Items table is the natural home for bag
  mutations even though they touch `currentZoneSession` as a side effect.
- `convex/_shared/character.ts` — private helpers shared across the three
  modules: `loadOwnedCharacter`, `loadEquippedSet`, `deleteZoneBag`,
  `fetchInventoryAllocator`, `newZoneSession`, `equippedSlotValidator`.
  Re-exporting from a `_shared` module is the cleanest way to avoid
  circular imports between characters / combat / items.

No behavior changes — pure code move + import rewrites. Schema, mutation
logic, error strings (so `translateServerError` keeps matching), and
tests are all untouched. The HP fix from the previous commit lands on
both `convex/characters.ts:create` and the moved combat mutations
unchanged.

Client call sites updated mechanically (~30 across world.tsx,
InventoryModal.tsx, useCombatLoop.ts): `api.characters.equipItem` →
`api.items.equipItem`, `api.characters.enterCity` → `api.combat.enterCity`,
etc. `api.characters.list`, `create`, `remove`, `byId` stay put
(character-select.tsx, CreateCharacterModal.tsx).

Validation:
- `npx convex dev --once` regenerates `_generated/api.d.ts` with the new
  `combat` and `items` namespaces.
- `npx tsc --noEmit` clean on both client and convex tsconfig.
- `npx vitest run`: 204/204 tests passing.
