+++ convex/_generated/api.d.ts
@@ -8,10 +8,13 @@
  * @module
  */
 
+import type * as _shared_character from "../_shared/character.js";
 import type * as auth from "../auth.js";
 import type * as characters from "../characters.js";
+import type * as combat from "../combat.js";
 import type * as http from "../http.js";
 import type * as itemValidator from "../itemValidator.js";
+import type * as items from "../items.js";
 import type * as users from "../users.js";
 
 import type {
@@ -21,10 +24,13 @@ import type {
 } from "convex/server";
 
 declare const fullApi: ApiFromModules<{
+  "_shared/character": typeof _shared_character;
   auth: typeof auth;
   characters: typeof characters;
+  combat: typeof combat;
   http: typeof http;
   itemValidator: typeof itemValidator;
+  items: typeof items;
   users: typeof users;
 }>;
